+++ allink_core/apps/config/abstract_models.py
@@ -47,146 +47,6 @@ class BaseConfig(TranslatableModel, SingletonModel):
         blank=True,
         null=True,
     )
-    news_verbose = models.CharField(
-        _(u'News verbose name'),
-        default=_(u'News entry'),
-        max_length=255
-    )
-    news_verbose_plural = models.CharField(
-        _(u'News verbose name plural'),
-        default=_(u'News'),
-        max_length=255
-    )
-    news_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
-    events_verbose = models.CharField(
-        _(u'Events verbose name'),
-        default=_(u'Event'),
-        max_length=255
-    )
-    events_verbose_plural = models.CharField(
-        _(u'Events verbose name plural'),
-        default=_(u'Events'),
-        max_length=255
-    )
-    events_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
-    eventsregistration_verbose = models.CharField(
-        _(u'Events Registration verbose name'),
-        default=_(u'Event Registration'),
-        max_length=255
-    )
-    eventsregistration_verbose_plural = models.CharField(
-        _(u'Events Registration verbose name plural'),
-        default=_(u'Event Registrations'),
-        max_length=255
-    )
-    eventsregistration_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
-    locations_verbose = models.CharField(
-        _(u'Locations verbose name'),
-        default=_(u'Location'),
-        max_length=255
-    )
-    locations_verbose_plural = models.CharField(
-        _(u'Locations verbose name plural'),
-        default=_(u'Locations'),
-        max_length=255
-    )
-    locations_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
-    people_verbose = models.CharField(
-        _(u'People verbose name'),
-        default=_(u'Person'),
-        max_length=255
-    )
-    people_verbose_plural = models.CharField(
-        _(u'People verbose name plural'),
-        default=_(u'People'),
-        max_length=255
-    )
-    people_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
-    testimonials_verbose = models.CharField(
-        _(u'Testimonials verbose name'),
-        default=_(u'Testimonial'),
-        max_length=255
-    )
-    testimonials_verbose_plural = models.CharField(
-        _(u'Testimonials verbose name plural'),
-        default=_(u'Testimonials'),
-        max_length=255
-    )
-    testimonials_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
-    work_verbose = models.CharField(
-        _(u'Work verbose name'),
-        default=_(u'Project/ Reference'),
-        max_length=255
-    )
-    work_verbose_plural = models.CharField(
-        _(u'Work verbose name plural'),
-        default=_(u'Projects/ References'),
-        max_length=255
-    )
-    work_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
-    members_verbose = models.CharField(
-        _(u'Members verbose name'),
-        default=_(u'Member'),
-        max_length=255
-    )
-    members_verbose_plural = models.CharField(
-        _(u'Members verbose name plural'),
-        default=_(u'Members'),
-        max_length=255
-    )
-    members_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
-    contactrequest_verbose = models.CharField(
-        _(u'Contact verbose name'),
-        default=_(u'Contact request'),
-        max_length=255
-    )
-    contactrequest_verbose_plural = models.CharField(
-        _(u'Contact verbose name plural'),
-        default=_(u'Contact requests'),
-        max_length=255
-    )
-    contactrequest_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
-    terms_verbose = models.CharField(
-        _(u'Terms of Service verbose name'),
-        default=_(u'Terms of Service'),
-        max_length=255
-    )
-    terms_verbose_plural = models.CharField(
-        _(u'Terms of Service verbose name plural'),
-        default=_(u'Terms of Service'),
-        max_length=255
-    )
-    terms_toolbar_enabled = models.BooleanField(
-        _(u'Toolbar enabled?'),
-        default=True
-    )
     config_allink_page_toolbar_enabled = models.BooleanField(
         _(u'allink Page Extension Toolbar Enabled?'),
         default=False
@@ -207,16 +67,16 @@ class BaseConfig(TranslatableModel, SingletonModel):
         default='itcrowd@allink.ch',
         max_length=255
     )
-    
+
     class Meta:
         abstract = True
         verbose_name = _(u'Configuration')
-    
+
     def __str__(self):
         return u'Configuration'
-        
+
         # SOLO_CACHE does not work in our setup, thats why, we rewrite it here
-    
+
     def to_dict(self):
         fields = {
             field.name: getattr(self, field.name) for field in self._meta.get_fields() if
@@ -238,27 +98,27 @@ class BaseConfig(TranslatableModel, SingletonModel):
             field.name not in ['master', 'language_code']
             })
         return fields
-    
+
     def set_to_cache(self):
         cache_key = self.get_cache_key()
         timeout = 60 * 60 * 24 * 180
         cache.set(cache_key, self.to_dict(), timeout)
-        
+
         # invalidate cache for favicon templatetag
         cache.delete('favicon_context')
-    
+
     def save(self, *args, **kwargs):
         if self.pk:
             self.set_to_cache()
         else:
             self.pk = 1
         super(SingletonModel, self).save(*args, **kwargs)
-    
+
     @classmethod
     def get_cache_key(cls):
         prefix = 'solo'
         return '%s:%s_%s' % (prefix, cls.__name__.lower(), get_language())
-    
+
     @classmethod
     def get_solo(cls):
         cache_key = cls.get_cache_key()
@@ -285,7 +145,7 @@ class BaseConfig(TranslatableModel, SingletonModel):
 
 class BaseConfigTranslation(TranslatedFieldsModel):
     master = models.ForeignKey('config.Config', related_name='translations', null=True)
-    
+
     default_base_title = models.CharField(
         verbose_name=_(u'Base title'),
         max_length=50,
@@ -294,7 +154,7 @@ class BaseConfigTranslation(TranslatedFieldsModel):
         blank=True,
         null=True
     )
-    
+
     class Meta:
         abstract = True
         app_label = 'config'
@@ -306,7 +166,7 @@ class BaseAllinkPageExtension(PageExtension):
         _(u'Special Subnav enabled?'),
         default=False
     )
-    
+
     class Meta:
         abstract = True
         app_label = 'config'