+++ allink_core/apps/contact/cms_toolbars.py
@@ -23,5 +23,4 @@ class ContactToolbar(CMSToolbar):
 
 
 
-if getattr(Config.get_solo(), 'contactrequest_toolbar_enabled', True):
-    toolbar_pool.register(ContactToolbar)
+toolbar_pool.register(ContactToolbar)