+++ allink_core/core/models/models.py
@@ -133,21 +133,11 @@ class AllinkBaseModel(models.Model):
 
     @classmethod
     def get_verbose_name(cls):
-        try:
-            Config = get_model('config', 'Config')
-            field_name = cls._meta.model_name + '_verbose'
-            return getattr(Config.get_solo(), field_name)
-        except:
-            return cls._meta.verbose_name
+        return cls._meta.verbose_name
 
     @classmethod
     def get_verbose_name_plural(cls):
-        try:
-            Config = get_model('config', 'Config')
-            field_name = cls._meta.model_name + '_verbose_plural'
-            return getattr(Config.get_solo(), field_name)
-        except:
-            return cls._meta.verbose_name_plural
+        return cls._meta.verbose_name_plural
 
     def is_published(self):
         return self in self.get_published()